Logistic Regression Model of Salifort Motors
An Logistic Regression Model of Salifort Models which predicts whether the employee of Salifort Motors will leave the company or not. And what are the main reason that employees are leaving in such a large amount.
This is a Capstone Project from coursera which I completed while pursuing Google Professional Advance Data Analytics Certification. I hope you'll enjoy the insights that are accomplished while working on this project.