Tableau Dashboard of Atliq Sales Insight
This project represents the visualization skills an Data Analyst should have. Here, I have divided Atliq Sales in smaller sub sales with respect to year and months. The stakeholder of Atliq as well as Sales team will have an descriptive idea regarding sales of the company and can make decisions for the company to gradually come back in business.
It is an dynamic dashboard where you can see the changes while applying filters on the top and can learn more about the sales of Atliq according to different parameters.